// src/enums/beans/beanFunctionPipe.ts
export enum BEAN_FUNCTION_PIPE_ENUM {
  IS_FROZEN = 'IS_FROZEN',
  IS_UNFROZEN = 'IS_UNFROZEN',
  IS_SCANNED_BEAN = 'IS_SCANNED_BEAN',
  HAS_PHOTOS = 'HAS_PHOTOS',
  HAS_FROZEN_INFORMATION = 'HAS_FROZEN_INFORMATION',
  BEAN_AGE_IN_DAYS = 'BEAN_AGE_IN_DAYS',
  HAS_CUSTOM_FLAVORS = 'HAS_CUSTOM_FLAVORS',
  HAS_PREDEFINED_FLAVORS = 'HAS_PREDEFINED_FLAVORS',
  IS_SELF_ROASTED = 'IS_SELF_ROASTED',
  SHOW_COST_PER_KG = 'SHOW_COST_PER_KG',
}